requirements.txt
flask
pandas
numpy
scikit-learn
xgboost
catboost
python-whois
requests
scipy
matplotlib
seaborn
